Fix homepage/footer +
Add services and sponsors db side

+++ database/migrations/2021_08_31_134414_create_sponsors_table.php
@@ -17,7 +17,7 @@ class CreateSponsorsTable extends Migration
             $table->id();
             $table->string('name');
             $table->decimal('price', 4, 2);
-            $table->tinyInteger('period');
+            $table->Integer('period');
             $table->timestamps();
         });
     }
